[EFL] Remove unused accessibility related code https://bugs.webkit.org/show_bug.cgi?id=153543
Reviewed by Darin Adler.

Source/WebKit2:

Do not need to load external library to expose WebKit's
accessibility tree. There have been changes in EFL/Elementary
in terms of support of accessibility and we should follow those.

* WebProcess/efl/WebProcessMainEfl.cpp:
(eailLibraryPath): Deleted.
(eail): Deleted.

Tools:

Removed reference to EAIL library.
It not going to be support anymore. There have been changes in
EFL/Elementary in terms of the accessibility and we should
follow those.

* efl/jhbuild-optional.modules:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@195725 268f45cc-cd09-0410-ab3c-d52691b4dbfc