Web Inspector: Option+Up/Down should not move cursor outside of number https://bugs.webkit.org/show_bug.cgi?id=153784 <rdar://problem/24453133>
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2016-02-01
Reviewed by Timothy Hatcher.

* UserInterface/Views/CodeMirrorAdditions.js:
When selectionStart === selectionEnd we were duplicating the
movement mutation by performing it twice on the same object.
After much experimentation, I left in the existing code path
for handling mutation with a selection. It is not perfect,
but it is better then just applying the ch diff.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@196003 268f45cc-cd09-0410-ab3c-d52691b4dbfc