<attachment> icon should be a folder for the custom MIME type multipart/x-folder https://bugs.webkit.org/show_bug.cgi?id=153795 <rdar://problem/24416632>
Reviewed by Anders Carlsson.

Source/WebCore:

Test: fast/attachment/attachment-folder-icon.html

* rendering/RenderThemeMac.mm:
(WebCore::iconForAttachment):
(WebCore::paintAttachmentIcon):
Mail uses this special MIME type to indicate that something is a folder, which there
isn't a normal non-deprecated MIME type for.

LayoutTests:

* fast/attachment/attachment-folder-icon-expected.html: Added.
* fast/attachment/attachment-folder-icon.html: Added.
Make sure that the rendering of an attachment with multipart/x-folder
matches that of a file reference pointing at a folder.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@196028 268f45cc-cd09-0410-ab3c-d52691b4dbfc