offlineasm: Emit Dwarf2 file and location directives to allow for debugging .asm files https://bugs.webkit.org/show_bug.cgi?id=152703
Reviewed by Mark Lam.

Added support to output Dwarf2 .file and .loc assembler directives to provide the debugging
information needed to correlate the offline assembler generated code with the source lines 
in the .asm files.

Changed the tracking of file data to include a file index that was provided to the .file
directive.  That index is used when emitting the .loc directives.

* offlineasm/arm.rb:
* offlineasm/arm64.rb:
* offlineasm/asm.rb:
* offlineasm/backends.rb:
* offlineasm/config.rb:
* offlineasm/parser.rb:
* offlineasm/x86.rb:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@196541 268f45cc-cd09-0410-ab3c-d52691b4dbfc