Use of inlined asm statements causes problems for -std=c99 builds. https://bugs.webkit.org/show_bug.cgi?id=154507
Reviewed by Dan Bernstein.

Source/bmalloc:

* bmalloc/BAssert.h:

Source/WTF:

WTF's Assertions.h may inadvertantly get included by other projects that are built
with -std=c99.  The use of the inlined asm statements with the keyword "asm" is
not recognized when the -std compiler flag is used.

https://gcc.gnu.org/onlinedocs/gcc/Extended-Asm.html says "When writing code that
can be compiled with -ansi and the various -std options, use __asm__ instead of
asm (see Alternate Keywords)."

So, to be a good citizen, we can change the use of "asm" in CRASH() to "__asm__"
so that we don't break the build of such other projects.

* wtf/Assertions.h:



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@196871 268f45cc-cd09-0410-ab3c-d52691b4dbfc