Calling importNode on shadow root causes a crash https://bugs.webkit.org/show_bug.cgi?id=154570
Reviewed by Anders Carlsson.

Source/WebCore:

The bug was caused by a missing check in cloneNode. Added cloneNodeForBindings to explicitly throw
an NotSupportedError when it's called on a shadow root. We don't clone shadow root when deep-cloning
the tree so we don't have to check that condition.

The behavior of cloneNode is specified at:
http://w3c.github.io/webcomponents/spec/shadow/#the-shadowroot-interface
(it current says we should throw DATA_CLONE_ERR but I have an spec bug filed at
https://github.com/w3c/webcomponents/issues/393)

The behavior of importNode and adoptNode are specified in DOM4 specification:
https://dom.spec.whatwg.org/#dom-document-importnode
https://dom.spec.whatwg.org/#dom-document-adoptnode

Tests: fast/shadow-dom/Document-prototype-adoptNode.html
       fast/shadow-dom/Document-prototype-importNode.html
       fast/shadow-dom/Node-prototype-cloneNode.html

* dom/Document.cpp:
(WebCore::Document::importNode): Throw NotSupportedError when importing a shadow root.
* dom/Node.cpp:
(WebCore::Node::cloneNodeForBindings): Added.
* dom/Node.h:
* dom/Node.idl: Use cloneNodeForBindings here.

LayoutTests:

Added W3C-style testharness tests for calling cloneNode on a shadow root.

Also added tests for adoptNode and importNode.

* fast/shadow-dom/Document-prototype-adoptNode-expected.txt: Added.
* fast/shadow-dom/Document-prototype-adoptNode.html: Added.
* fast/shadow-dom/Document-prototype-importNode-expected.txt: Added.
* fast/shadow-dom/Document-prototype-importNode.html: Added.
* fast/shadow-dom/Node-prototype-cloneNode-expected.txt: Added.
* fast/shadow-dom/Node-prototype-cloneNode.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@196998 268f45cc-cd09-0410-ab3c-d52691b4dbfc